outsource input logic to input.ts. default to this.alpha instead of 1 when changing rgb value
