+++ README.md
@@ -3,3 +3,19 @@
 Core package for iro.js (WIP)
 
 This is for functionality that can be shared between the standalone version of iro.js as well as the upcoming React and Vue components.
+
+
+CHANGES: 
+========
+
+repo: https://github.com/lenadax/iro.js/tree/feature/slider-label-value
+
+- allow alpha value when setting r/g/b or h/s/l channels (color.ts)
+  - default to this.alpha instead of 1 when setting rgb value
+
+- logic for editable input field (input.ts)
+
+TODO:
+=====
+
+[ ] - allow seperate slider width spec
