+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './color';
 export * from './slider';
+export * from './input';
 export * from './wheel';
 export * from './box';
 export * from './util';
